Part 1 - contentPage.tsx uses new Post component
* Separate out the Post/Header/Footer to diff components
* Post header deals with creator info & edit/delete content
* Post footer deals with up/downvote content
* Use a context provider to more easily pass information between the dependencies

Next steps:
* Integrate into the new component into the singleton view
* Support rendering context - i.e. we do and need different things when rendering
  on the feed page vs the singleton page.
